[nextest] add fail-fast support
Add support for canceling test runs if a test fails.

The current UI for mutually exclusive options sucks, so hide it behind a
setter interface.

Also switch to using a nextest config for the test fixture, which found
a bunch of issues in the config definition -- fix them up too.