[release] update releases.json, fix publish action
The action I was using appears to parse git's porcelain output, which
isn't suitable for scripts and was causing issues for our setup. Replace
it with just calling git directly.

This is still potentially racy between the git pull and the push but the
throughput of this repo isn't high enough for this to be a major concern.