[nextest-runner] switch to AND semantics
Previously, if either expression OR string filters matched, we'd run the tests.
The current (stable) implementation is that *both* the expression AND the string
filters should match.

This is because we try and skip running test binaries which don't match
expression filters. So for example:

    cargo nextest run -E 'binary(foo)' test_bar

would not even get to the point of enumerating the tests not in binary(foo), thus
not running any test_bars in the workspace. But, with the OR semantics:

    cargo nextest run -E 'binary(foo) or test(test_foo)' test_bar

would run all the test_bars in the repo. This is inconsistent.