[nextest-runner] put each test into a process group on unix
This allows for a process and all its child processes to be signaled
atomically.

A child process spawned by a test can choose to run in another process
group, but that's quite uncommon.

The one regression here is that if you press Ctrl-C, tests no longer
exit. In upcoming commits we're going to fix those regressions by
forwarding Ctrl-C to child processes.