fix: respect test runner during build reuse
Previously `CARGO_TARGET_<target>_RUNNER` would be ignored if a build was
loaded from an archive file with --archive, because no/the incorrect target-triple was
detected in that case (see #468).

This patch changes the runner detection to account for those cases.
The target triple is now stored in the `BuildMetadata`.
If cargo is invoked the triple will be detected with the usual mechanism while
the `BuildMetadata` is generated.
During build-reuse the target triple is loaded from the `BuildMetadataSummary`.

If the target was not manually set, the host target-triple will be used
during serialization. This ensures correct behaviour in cases where
runner has a different default toolchain.