Optimization: Collect output in parallel in `TestRunner::run_test_inner`
* Add new dep tokio-util v0.7.4 to nextest-runner
* Create new type `ReadTask` for spawning the read tasks and handles the
  cancellation.

This moves collection of output to a separate tokio task instead of
doing it in the `tokio::select!` loop, which can be the bottleneck since
`tokio::select!` is quite complex.

This also simplify the implementation by replacing the second
`tokio::select!` loop with a spawned job.

All of the `*_done` boolean flags is also removed.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>