Strip ANSI escapes from JUnit output (#1054)
Previously the escape character itself was removed, but the rest of the escape sequence was left in place, resulting in useless and seemingly random characters in the output. (Unless some tools expect this and know how to handle it, but that isn't the case for the tools I use.)

There was already a test that asserted this. I assumed the goal of that test was just to validate that illegal characters (like the escape character) didn't make it into the output, and I've adjusted it to verify that the entire sequence is stripped out.