Add support for combined test output (#1212)
For libtest JSON emulation, we need to have a way to model combined stdout and stderr output. The best way to do this happens to be to use the same file descriptor (HANDLE on Windows) for stdout and stderr.

This is pretty useful generally, and in the future we may enable support for combined test output even outside the libtest-json context.

The main alternative considered was having separate fds, but selecting from them into a single giant buffer with tracking indexes into it. However, that doesn't quite work because reading from pipes is done asynchronously with the test process writing to them. This typically causes lines from stdout and stderr to be clumped with each other.

If the same fd is used as in this PR, the test process writes to stdout and stderr *synchronously*, which means that the combined output is recorded in the exact order the test process generates it in.

Enabling combined output does mean that we lose the ability to track outputs separately, but that's not a huge deal.

Co-authored-by: Rain <rain@sunshowers.io>