[nextest-filtering] Simplify transition to 0.5 with 'unpeek'
In 0.5, `parse_next` modifies input so the old behavior is under the
name `parse_peek` (as it doesn't commit to the parse but returns what it
would commit).
So `unpeek` takes a combinator that doesn't modify input ("peek parses")
and makes it commit to the result modifying the input.

This is added only to ease the transition.  Future commits will work to
remove `unpeek` calls.