[nextest-runner] add a `default-set` config option
It can often be useful to have a subset of tests skipped by default. In order
to make that happen, add the ability to configure a `default-set` config
option. This is `all()` by default.

A worry with any such configuration is users complaining about missing tests.
To ensure that people notice that, we also bubble this up quite prominently in
the UI.