[nextest-runner] highlight interesting parts of the output
While looking at long outputs, it can sometimes be hard to find the exact point
of failure. With this change, the 2 lines that we think are most relevant to
the failure are highlighted.

This is heuristic so we can get it wrong -- we don't filter the output, just
highlight the lines we think are most important.

We do use some unsafe code to compute the right indexes. This is temporary and
the next commit will remove it.