Add support for silencing only one of the outputs of a test
The old syntax is still supported, the new complete syntax is:
 - "{value}":                         where value is one of [immediate, immediate-final, final, never]
 - "{stream}={value}":                where {stream} is one of [stdout, stderr]
 - "{stream}={value},{stream=value}": where the two {stream} keys cannot be the same

For tests where the output is combined (currently only used for
libtest-json output), only the stdout value is used.

There is a small regression where the CLI won't suggest values that
are close to the incorrect input, i.e. if you input 'imediate' it won't
suggest 'immediate'.