[cargo-nextest] make version computation less expensive
Realized that os_info would do a ton of calls that aren't necessary, slowing
down nextest quite measurably. It's hard to integrate lazy version evaluation
into clap (https://github.com/clap-rs/clap/issues/5543) so let's just not print
OS info for now.

If we do want OS info in the future it'll probably be as part of a support
tarball or similar.