[nextest-runner] centralize channel management in CallbackContext
In upcoming work we're going to add the ability to query test tasks. For a
coherent UX it's important that what the context reports is consistent with
tests. (For example, if the displayer hasn't shown a test as started, it is
surprising to see output for it.)

In order to achieve that, move away from broadcast channels towards
centralizing all channel management code in `CallbackContext`. With this
change, execution units now ensure that the callback context is synced up with
them before proceeding.