[nextest-runner] make the runner loop own the test output buffers
Currently, when collecting output from tests, the corresponding futures own the
buffers into which output is being collected. While this works for simpler use
cases, we'd like to be able to dump in-progress output to the terminal in the
future.

To enable this use case, use the "externalized progress" pattern pioneered by
APIs like Tokio's `AsyncReadExt::read_buf`: maintain an external set of buffers
to accumulate output into called `ChildOutputMut`, and write into them within
the corresponding select loops.

In the future, on receiving an instruction to get the current standard
output/standard error, the select statement would be able to look at these
buffers and return whatever's currently in them.

Also do some general cleanup of comments and error reporting.