[nextest-runner] simplify reporter error handling
Trying to handle reporter errors in the main runner loop is a bit of a mess.
Instead, handle them at a higher level and use a oneshot channel to indicate
that an error has occurred. This fits in quite naturally with the rest of the
event-driven system.

Also change "canceled" to "cancelled" for consistency.