[nextest-runner] handle outputs with/without trailing newlines better
Previously, our headers would always insert a leading newline to compensate for
situations where the output didn't have a trailing newline in it. This isn't
quite right (what if the last output doesn't have a newline in it)? It also
causes issues with situations where RESET_COLOR needs to be inserted
immediately before the last newline.

Change our code to always insert a trailing newline, and the trailer
immediately before that.