[nextest-runner] handle SIGINFO and SIGUSR1 to dump info to screen (#1938)
Part 1 of work to perform interactive queries.

With this change, if nextest receives either SIGINFO (where available)
or SIGUSR1, it will print a list of all currently running tests with
their states. The full test state machine is modeled in the event
responses, along with possible sources of errors.

I tried making Ctrl-Break on Windows do the same thing, but
unfortunately that doesn't quite work in practice. That's because
Ctrl-Break is received by all processes connected to the console, not
just by the equivalent of the "foreground process group". The printing
out of info works, but running tests are immediately aborted as well. So
remove this. (In interactive terminals we'll support an alternative --
pressing `i` -- which should work for most use cases on Windows.)

There is also an undocumented environment variable
`__NEXTEST_SIGQUIT_AS_INFO` which allows `SIGQUIT` (`Ctrl-\`) to perform
an info query. Quite useful for testing on Linux, where SIGINFO is sadly
unavailable.

Note: illumos does support SIGINFO, but that is blocked on an upstream
issue: https://github.com/tokio-rs/tokio/pull/6995