[nextest-runner] support pressing `t` to dump current test info (#1911)
In prior work (#1938) we enabled the `SIGUSR1` and `SIGINFO` handlers to
dump internal state to standard output.

However, those handlers aren't present on all platforms.

* `SIGINFO` doesn't exist on Linux. `SIGUSR1` exists but is hard to send
signals over (have to use `kill`, nothing like `^T`).
* Windows' very basic signal handling mechanism isn't really up to the
task.

As an alternative, allow these platforms to specify `t` to dump current
test information to the interactive terminal.

It would also be interesting to support a Unix domain socket as a way to
query current state in the future, and have a separate instance of
nextest be able to connect to the socket -- this work naturally builds
towards that kind of model.