[nextest-runner] use a setup script to seed build for integration tests
Currently, all our integration tests build a separate copy of the nextest
fixtures. This is a lot of unnecessary extra build work.

We now have a great solution, though -- use setup scripts! This kind of build
caching is exactly what setup scripts are good at, and this also gives us a
chance to dogfood them.

This is a really nice performance improvement on my local machine (goes from
~5s to ~3s) -- should see a similar or even better result in CI.