[nextest-runner] use a setup script to seed build for integration tests (#1963)
Currently, all our integration tests build a separate copy of the nextest
fixtures. This is a lot of unnecessary extra build work.

We now have a great solution, though -- use setup scripts! This kind of build
caching is exactly what setup scripts are good at, and this also gives us a
chance to dogfood them in the nextest repo proper (rather than just in the
fixtures).

This is a really nice performance improvement on my local machine (goes from
~5s to ~3s) -- should see a similar or even better result in CI.

To make the tests still work, I had to do two internal/unstable things:

1. Add a `debug current-exe` command to get the current executable. Not
part of the stable interface, though not hugely painful to stabilize it
if there's a compelling use case.
2. Add `__NEXTEST_ALT_TARGET_DIR` to ensure that we're still testing the
situation where linked paths aren't found. This is definitely a
test-only knob that shouldn't be stabilized.