[nextest-runner] spawn a task for each test
Previously we'd use the same task for all tests -- that turns out to be fine in
most cases, but doesn't quite provide per-test isolation, and is slower in
repos like clap that have lots of small tests. Switch to using the same task
instead.

It's cool that async-scoped kind of works here, but non-static tasks are such
a pain to work with in general.

This makes the clap tests significantly faster. On `fc55ad08`, runtime on my
desktop (Ryzen 7950X/Linux) goes from 0.36 seconds to 0.24.