[nextest-runner] remove cancellation AtomicBool
We've changed things such that the executor always checks in with the
dispatcher before starting a unit. This makes the bool unnecessary.

The cancel receiver is currently necessary for the
handle_delay_between_attempts step, which should cancel the run no matter what.
We'll remove this shortly by moving this functionality into req_rx.