[nextest-runner] on Unix, check that nextest controls the terminal (#2283)
There's a difference between stdin being a terminal and nextest actually
being in control of the terminal. Also do the latter check on Unix.

On Windows, this doesn't appear to be necessary, since it doesn't have a notion of foreground process groups.

This makes watchexec + nextest no longer hang. I tested and watchexec +
nextest don't hang on Windows either.