[nextest-runner] correctly dup the fd with combined output (#2316)
## The bug

Previously, for combined output, we were using the same file descriptor
and storing it three times:

1. In `.stdout`
2. In `.stderr`
3. In `state.theirs`.

`Stdio` consumes ownership of the fd, so this effectively meant that
three `OwnedFd` instances were present, and fds would be dropped three
times.

Now in a single-threaded program this is relatively harmless -- Rust
just gets `EBADF` for the latter two calls. But in multithreaded
programs this was completely broken, with a bit of a horrific race
condition on Linux. Since Tokio uses pidfds for process event handling
on Linux, this caused missed wakeups in this scenario:

1. Let's say we allocated fd 20 as the `theirs` end provided to the
child process `Stdio`.
2. In thread 1, we dropped the child, closing fd 20 the first time.
3. In thread 2, Tokio created a new pidfd and the kernel allocated fd 20
to this pidfd.
4. In thread 1, we continued cleaning up the child, closing fd 20 two
more times.
5. The second fd 20 close caused epoll to no longer dispatch events to
the pidfd.
6. After the test process exited, nextest never got a notification about
that happening.

## The fix

To fix the bug:

1. Don't store `theirs` in `State` -- we don't need it, and in fact had
an `expect(dead_code)` on it.
2. Call `try_clone` to ensure that stdout and stderr get two separate fd
numbers that refer to the same pipe.

## Notes

* This wasn't originally visible because we used a single task for all
threads (so there was no multithreaded race), but 5f010c7 caused each
test to be in its own task. This change exposed the bug.
* This is particularly visible with pidfd because it uses the same fd
namespace as regular descriptors. With more traditional SIGCHLD-based
handling, the bug wouldn't be as visible.

Fixes #2295.