Add a counter to the status line with each test (#2618)
This adds a counter to each status line, so that one doesn't have to
guess how many tests are remaining when the progress bar is not visible
(think in CI):

It's been a long requested feature: #526. As a casual observer of test
runs with 5000+ tests, this would be nice to have.